refactor(dropzone-image): extract answer zone resizing into a hook
