+++ packages/editor/src/plugins/dropzone-image/components/answer-zone/answer-zone.tsx
@@ -1,6 +1,6 @@
-import { useContext, useEffect, useState } from 'react'
+import { useContext } from 'react'
 import { useDrag } from 'react-dnd'
-import { ResizableBox, ResizableBoxProps } from 'react-resizable'
+import { ResizableBox } from 'react-resizable'
 
 // eslint-disable-next-line import/no-unassigned-import
 import 'react-resizable/css/styles.css'
@@ -9,18 +9,15 @@ import { AnswerZoneAnswer } from './answer-zone-answer'
 import { AnswerZoneEmpty } from './answer-zone-empty'
 import { AnswerZoneSidebar } from './answer-zone-sidebar'
 import { AnswerZonesContext } from '../../context/context'
-import { AnswerType, AnswerZoneState, DropzoneVisibility } from '../../types'
-import { getPercentageRounded } from '../../utils/percentage'
+import { useAnswerZoneResize } from '../../hooks/use-answer-zone-resize'
+import {
+  type AnswerType,
+  type AnswerZoneState,
+  DropzoneVisibility,
+} from '../../types'
 import { cn } from '@/helper/cn'
 
 export const answerZoneDragType = 'answerZone'
-
-interface PositionState {
-  left: number
-  top: number
-  width: number
-  height: number
-}
 
 export interface AnswerZoneProps {
   answerZone: AnswerZoneState
@@ -42,29 +39,13 @@ export const AnswerZone = (props: AnswerZoneProps) => {
     onClickPlusButton,
     onClickEditAnswerButton,
   } = props
-  const left = answerZone.position.left.get()
-  const top = answerZone.position.top.get()
-  const height = answerZone.layout.height.get()
-  const width = answerZone.layout.width.get()
   const name = answerZone.name.get()
-
-  const [positionState, setPositionState] = useState<PositionState>({
-    left: canvasWidth * left,
-    top: canvasHeight * top,
-    width: canvasWidth * width,
-    height: canvasHeight * height,
-  })
-  useEffect(() => {
-    setPositionState({
-      left: canvasWidth * left,
-      top: canvasHeight * top,
-      width: canvasWidth * width,
-      height: canvasHeight * height,
-    })
-  }, [canvasHeight, canvasWidth, height, left, top, width])
 
   const context = useContext(AnswerZonesContext)
   const { dropzoneVisibility } = context || {}
+
+  const { positionState, minWidth, minHeight, handleResize, handleResizeStop } =
+    useAnswerZoneResize({ answerZone, canvasHeight, canvasWidth })
 
   const [collected, drag, dragPreview] = useDrag({
     type: answerZoneDragType,
@@ -74,86 +55,6 @@ export const AnswerZone = (props: AnswerZoneProps) => {
       hideSourceOnDrag: true,
     }),
   })
-
-  const handleResize: ResizableBoxProps['onResize'] = (_, { size, handle }) => {
-    setPositionState((previous) => ({
-      ...size,
-      left:
-        handle === 'nw' || handle === 'sw'
-          ? previous.left + previous.width - size.width
-          : previous.left,
-      top:
-        handle === 'nw' || handle === 'ne'
-          ? previous.top + previous.height - size.height
-          : previous.top,
-    }))
-  }
-
-  const handleResizeStop: ResizableBoxProps['onResizeStop'] = (
-    _,
-    { size, handle }
-  ) => {
-    if (handle === 'nw' || handle === 'sw') {
-      const leftOnResizeStop =
-        positionState.left + positionState.width - size.width
-      const isOverflowingLeft = leftOnResizeStop < 0
-      const newLeft = isOverflowingLeft ? 0 : leftOnResizeStop
-      const newWidth = isOverflowingLeft
-        ? positionState.width + leftOnResizeStop
-        : size.width
-      answerZone.position.left.set(getPercentageRounded(canvasWidth, newLeft))
-      answerZone.layout.width.set(getPercentageRounded(canvasWidth, newWidth))
-    } else {
-      const rightOnResizeStop = positionState.left + size.width
-      const isOverflowingRight = rightOnResizeStop > canvasWidth
-      const newWidth = isOverflowingRight
-        ? positionState.width + (canvasWidth - rightOnResizeStop)
-        : size.width
-      answerZone.layout.width.set(getPercentageRounded(canvasWidth, newWidth))
-    }
-
-    if (handle === 'nw' || handle === 'ne') {
-      const topOnResizeStop =
-        positionState.top + positionState.height - size.height
-      const isOverflowingTop = topOnResizeStop < 0
-      const newTop = isOverflowingTop ? 0 : topOnResizeStop
-      const newHeight = isOverflowingTop
-        ? positionState.height + topOnResizeStop
-        : size.height
-      answerZone.position.top.set(getPercentageRounded(canvasHeight, newTop))
-      answerZone.layout.height.set(
-        getPercentageRounded(canvasHeight, newHeight)
-      )
-    } else {
-      const bottomOnResizeStop = positionState.top + size.height
-      const isOverflowingBottom = bottomOnResizeStop > canvasHeight
-      const newHeight = isOverflowingBottom
-        ? positionState.height + (canvasHeight - bottomOnResizeStop)
-        : size.height
-      answerZone.layout.height.set(
-        getPercentageRounded(canvasHeight, newHeight)
-      )
-    }
-  }
-
-  const minWidth = Math.max(canvasHeight * 0.2, 110)
-  const minHeight = Math.max(canvasHeight * 0.09, 45)
-
-  useEffect(() => {
-    setPositionState((previous) => ({
-      ...previous,
-      width: positionState.width < minWidth ? minWidth : previous.width,
-      height: positionState.width < minWidth ? minHeight : previous.height,
-    }))
-    if (positionState.width < minWidth) {
-      answerZone.layout.width.set(minWidth / canvasWidth)
-    }
-    if (positionState.height < minHeight) {
-      answerZone.layout.height.set(minHeight / canvasHeight)
-    }
-    // Only check once
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
 
   // Hide source element while dragging
   if (collected.isDragging) {
